LayoutTests/fast/css/rotate-invalidate-if-disabled.html

<!DOCTYPE html><!-- webkit-test-runner [ CSSIndividualTransformPropertiesEnabled=false ] -->
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test.js"></script>
</head>
<body>
<div id="target" style="display: none"></div>
<script>
description("Tests that rotate is not exposed when the feature is disabled");

shouldBeFalse("'rotate' in document.documentElement.style");
shouldBeFalse("'rotate' in getComputedStyle(document.documentElement)");
shouldBeTrue("getComputedStyle(document.getElementById('target')).getPropertyValue('rotate') == ''");
shouldBeFalse("CSS.supports('rotate: contain')");
shouldBeFalse("CSS.supports('rotate: inherit')");
shouldBeFalse("CSS.supports('rotate', 'inherit')");

</script>
</body>
</html>
