LayoutTests/fast/css/rotate-invalidate-if-disabled-expected.txt

Tests that rotate is not exposed when the feature is disabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'rotate' in document.documentElement.style is false
PASS 'rotate' in getComputedStyle(document.documentElement) is false
PASS getComputedStyle(document.getElementById('target')).getPropertyValue('rotate') == '' is true
PASS CSS.supports('rotate: contain') is false
PASS CSS.supports('rotate: inherit') is false
PASS CSS.supports('rotate', 'inherit') is false
PASS successfullyParsed is true

TEST COMPLETE
